benchmarks/bm_pytorch_alexnet_inference/requirements.txt
future==0.18.2
numpy==1.19.0
Pillow==8.0.0
torch==1.5.1
torchvision==0.6.1
